Add extraction modes `all`, `unique`, and `first`
Mode `all` returns all matches that are found for the given combination
of haystack and needle.

Mode `unique` depuplicates the matches that are found for the given
combination of haystack and needle.

Mode `first` returns only the first match that is found.

All modes return an empty list if no matches are found.